Add support for session daemon auto spawn
The lttng tool now spawns a session daemon if no ltt-sessiond is
running. Two new options added to control this behavior.

Introduce --sessiond-path and --no-sessiond to lttng command line tool.

Add --sig-parent to ltt-sessiond that sends a SIGCHLD to the parent pid
when the daemon is ready to process client command.

Signed-off-by: David Goulet <david.goulet@polymtl.ca>