Major changes for the daemon data transmission
This commit reengineer the data transmission from the session daemon to
the liblttngctl. Instead of sending a packet on a per data structure
basis, this scheme uses lttcomm_lttng_msg structure has a header control
structure adding the payload_size to it. The payload data is then
appended after and only one send() is used.

This will permit AF_INET socket scaling for future development.

List sessions and list apps command has been modified to fit this
change.

In the meantime, some unuse code is removed and minor cleanup.

Signed-off-by: David Goulet <david.goulet@polymtl.ca>