Add trace listing feature
This patch adds the possibility to list the available traces for a
specific session.

The new lttng_trace data structure is a representation of a trace for
use for 'human output'.

Some fct are added to trace.c to get the trace count of a session and
list those traces. Also, trace counter are added to a ltt_session.

Some renaming is done to make things clearer. A name is added to a
userspace trace also.

Signed-off-by: David Goulet <david.goulet@polymtl.ca>