Add support for shorten session uuid
Session are now identified by name.short_uuid where short_uuid is the
first 8 bytes of the 36 bytes uuid.

Also, fix the bug where you could'nt pass two command to the session
daemon from liblttngctl. The connect and disconnect is added to the
ask_sessiond function. So, on every command, a connect and close is
done.

Signed-off-by: David Goulet <david.goulet@polymtl.ca>