Add support for auto session creation
Support for auto session creation when no session (-s) is specified. The
session is created with the name being the current date and time. For
all commands that do not needs a session, no auto session is created and
for thoses who needs -s also.

Improve shorten uuid management all around the lttng client code.

Fix uuid usage in lttng_session data structure and a small fix in the
liblttngctl for the set_current_uuid fct.

Fix session name regex of the tests.

Signed-off-by: David Goulet <david.goulet@polymtl.ca>