Major changes on the lttng command line tool
Add create and destroy trace action but NOT implemented on the session
daemon side.

Add kernel create channel option but NOT implemented on the session
daemon side.

Add auto_session and auto_trace support for the lttng tool. This means
that if only --start is specified with a certain pid for instance, a
session is automatically created and a trace.

Fix a bug on the session daemon side where the ret value was changed
when no lttcomm_lttng_msg structure was allocated before replying to the
client.

Add a lot more command line validation for conflicting command.

Remove the use of UUID in the session name and for the create trace of
the liblttngctl API.

Remove the memset(0) applied to the llm structure of liblttngctl thus
resetting every information. This is not wanted for multiple command for
one lttng execution.

Signed-off-by: David Goulet <david.goulet@polymtl.ca>