Major changes
ltt-sessiond
-------------

Change command type to add the LTTNG_ prefix. Only kernel command at
this point working with the new kernel tracer lttng 2.0.

Adds support for the default channel.

Kernel tracer session, metadata and streams opening are now handled by
the session daemon by removing the public commands for those options.
This means that the public API of lttng-tools DO NOT permit to control
these three actions anymore.

Adds ulimit setting for open files (65535) for sessiond running as
UID=0.

lttng.h
-------------

As mention before, create_session, create_stream and open_metadata is
removed from the public API.

UUIDs are not exposed anymore by the API. Everything is done with the
session name.

New lttng event structure added in order to support kernel tracepoints,
kprobes and function tracer.

New lttng channel structure to create a kernel channel. Add the lttng
channel attribute structure also where to put the channel attributes.

libkernelctl
-------------

Adds flush buffer function and wait quiescent for kernel tracer.

Uses the new enable/disable kernel ioctl that replaces create event.

lttng
-------------

*MAJOR*: This is the inital import of the git-alike UI.

Some features are still not implemented but all the basics and more are
there.

This commit introduce a new feature where the session name is kept in
the current directory where the create command was executed (or
specified by the user with --path). A .lttng directory is created and a
config file that contains, for now, the session name. So, in order to
apply command on a session, you either specify it on the command line or
go in the directory where your .lttng dir reside.

A lot of code was changed/added/removed from all over the lttng-tools
tree to fit this big change. Consider this commit to be an almost
complete rewrite of the original git command line UI and a big part of
the public API.

Signed-off-by: David Goulet <david.goulet@polymtl.ca>