Add detection of debugfs and kernel module loading
The session daemon now detects the path of the debugfs mount point or
create and mount it at a default location (/mnt/debugfs).

Then, the kernel module are loaded automatically, if present, and the
kernel tracer fd is opened at the previous detected debugfs path.

The kernel module list is hardcoded for now.

NOTE: There is no module unload at the session daemon cleanup. This is
to avoid removing tracing capabilities when a trace is ongoing and
written to disk.

Signed-off-by: David Goulet <david.goulet@polymtl.ca>