Remove libustctl and libustcomm
Those libs are part of the soon to be deprecated user-space tracer. The
code is removed but not the directories and Makefile.am.

lttng-tools will only support UST 2.0 which is not yet released at this
point in time. Thus, this is not useful anymore and will be readded
later on to support the new user-space tracer.

Also disable a small portion of code in the session daemon that uses
libustctl. This code however remains relevant for future UST
integration.

Signed-off-by: David Goulet <david.goulet@polymtl.ca>