New public API for lttng control
Introduce the lttng domain that is used to identify the tracer.

Remove all 'kernel' from the functions name and now take the domain
struct as first parameter.

Add channel feature is removed and merged with enable-channel.

Move kernel specific data structure from lttng.h to lttng-kernel.h which
is copied from the LTTng kernel modules ABI. Genereic structures are
used to create a generic view of lttng event and attributes (agnostif of
the Kernel and user-space).

Default channel is NOT created unless it is needed.

A lot of comments has been added to lttng.h

Fix a bug in the liblttngctl. A return value was not set correctly.

Signed-off-by: David Goulet <david.goulet@polymtl.ca>