ltt-sessiond: use short options for modprobe in order to support Busybox
Many embedded systems are based on Busybox, and therefore use the
Busybox implementation of modprobe. This implementation does not
support long options such as --remove and --quiet, only short options
such as -r and -q are supported.

This patches changes ltt-sessiond to use the short options, which are
more widely available, and allows lttng to work easily on a
Busybox-based system.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>