Add new API call lttng_channel_set_default_attr
This call takes a lttng domain and an allocated channel attributes
pointer and set the default attributes of the channel according to the
domain type.

This commit also fix a listing problem where the wrong output type of
the channel was returned for UST.

The enable-channel command is also changed to used that new API call.

Signed-off-by: David Goulet <david.goulet@polymtl.ca>