Fix enable-event for UST events
When enabling an existing event, a new event creation occured which was
simply the bad thing to do. This patch fix that by adding the
enable_ust_event call for the case.

Also fixes a segfault where any UST object could be release with the
data pointer being set to NULL.

This patch introduce also assert() in the UST app code for
create/enable/disable event. The purpose is to catch, in dev, the bad
code execution flow and fix it.

Signed-off-by: David Goulet <david.goulet@polymtl.ca>