Add non-root support for UST tracing
User-space tracing can now be used without root access meaning that the
session daemon can run with the user credentials and trace his own
applications.

This adds the LTTNG_HOME_RUNDIR which is located, under non privileged
user, in $HOME/.lttng. It contains all the control and consumers
sockets. It is cleanup when lttng-sessiond dies.

Signed-off-by: David Goulet <david.goulet@polymtl.ca>