Complete change of the source directory tree
For the sake of clearness and awesomeness :P, this change was needed for
better semantic across the code and better organization of source files.

src/
	--> Contains source files of the project
	bin/
	--> Contains every binaries this git tree builds (lttng,
	lttng-sessiond and lttng-consumerd).

	lib/
	--> Contains exported libraries (lttng-ctl).

	common/
	--> Contains shared code for the entire project. (internal
	libraries mostly)

Every internal headers are moved from include/ to src/common. The
include directory only contains exported public headers now
(lttng/lttng.h). The -I$(top_srcdir)/src is also added for all
Makefiles.

The consumerd APIs are not exported anymore. The reason is that the
consumer will get completely re-engineered in the next months. It's our
safety net for the APIs breakage of the coming year concerning the
consumerd library.

Renaming is done also on some internal libraries to follow the naming
convention. The public library liblttngctl is renamed to liblttng-ctl
making it linkable now with "-llttng-ctl" deprecating the old
"-llttngctl".

No code is changed with this commit except #include statements
added/removed/renamed. This is why there is yet files and internal API
calls that does not follow the naming convention (Ex: lttngerr.h,
lttng-share.h in src/common).

Signed-off-by: David Goulet <dgoulet@efficios.com>