Check kernel version for tests
Test using the session daemon must be done on a kernel version 2.6.27 or
higher for lttng-tools and lttng-ust requirements. Before starting a
session daemon, tests validate the kernel version and stop them if the
kernel is too old.

This is *really* important for launchpad build system which compile and
run our tests on 2.6.24 kernel.

Signed-off-by: David Goulet <dgoulet@efficios.com>