Add CMD_WARNING error code
The CMD_WARNING is returned if at least one command went wrong and at
least one succeeded. This shoudl tell the user that there is an error
message on stderr but the rest of the command went well.

One use case for that is the multiple -t of add_context. If one of them
fails, this code is returned.

This error code is added at the end of the enum so the previous CMD code
are not changed.

Also, if the tracer (-k/-u) is not specified, CMD_ERROR is returned
instead of CMD_UNDEFINED which makes way more sense.

Signed-off-by: David Goulet <dgoulet@efficios.com>