Revert "Update loglevel ABI"
Segfault on lttng enable-event when no loglevel is specified.

UST tests failed.

This implementation forces a loglevel to each event which is fine
however the default level is never set when no loglevel is specified.

The internal UST ABI/API (used when lttng-ust is disabled) was not
updated also making it fail when disablin the lttng-ust support.

This broke too much stuff in lttng-tools so the commit is reverted and
test must pass before commiting it again.

lttng-ust HEAD will *NOT* work with lttng-tools head so stay on commit
939950af98d044a3b0632cad5293f91cb1493ef8 for UST.

This reverts commit dad47fc4c65ec0acbfedc676895833b465d25cd6.

Signed-off-by: David Goulet <dgoulet@efficios.com>