Improve error handling of lttng cli
Most of the commands were not returning a valid error code to the user.

Add "UST invalid context" lttcomm code and enforce "-t TYPE" for
add-context command.

Change error message for enable/disable event/channel in order to make
it standard for those commands and more meaningful to the user.

Signed-off-by: David Goulet <dgoulet@efficios.com>