Fix EEXIST handling on epoll_ctl add
Creating a second channel on the kernel tracer triggers a
update_kernel_poll which iterates over all channels and add the file
descriptors to the epoll set. Collision occurs with the first previously
added channel so EEXIST is acceptable and we simply do not perror() to
keep the session daemon stderr clean.

Signed-off-by: David Goulet <dgoulet@efficios.com>