Add the lttng view command
This adds "view" to the lttng command line interface. It's used to
visualize traces with a given viewers (default: babeltrace).

The -e, --viewer options is available to override the default viewer
settings or add arguments to the viewer.

For example:

$ lttng view -e "babeltrace -n scope --no-delta --clock-raw"

The trace directory path is automatically appended using the current
session or the given session name to lttng view.

The default behavior is to simply spawn babeltrace with "-n all".

$ lttng view
[...]

LTTv support is now disabled but is planned for the stable release.

Signed-off-by: David Goulet <dgoulet@efficios.com>