Fix session syscall listing
This problem occured because the "syscall" event was not added to the
event list since his kernel FD is not valid (fd=0). However, this fd is
only usefull if we want to add context or disable events. Those two
operations are not supported for now.

So, we can safely add the event to the list and add a hardcoded name in
the lttng cli list command.

Signed-off-by: David Goulet <dgoulet@efficios.com>