Fix: problem in port assignment in uri_parse()
The functions now puts 0 in the port field of the lttng_uri structure if
NO port are specified. It's now the caller's job to assign a default
port for that case.

Since the uri_parse() function does not know the URI context, a default
port can not be assigned correctly.

Signed-off-by: David Goulet <dgoulet@efficios.com>